requirements.txt
python-docx
python-dotenv
openai==1.58.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
faiss-cpu==1.9.0.post1
python-telegram-bot==20.7
